moving around trait stuff a bit more
