+++ src/griptape_nodes/exe_types/core_types.py
@@ -6,7 +6,6 @@ from copy import deepcopy
 from dataclasses import dataclass, field
 from enum import Enum, auto
 from typing import TYPE_CHECKING, Any, ClassVar, Self, TypeVar
-from griptape.mixins.singleton_mixin import SingletonMixin
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -655,27 +654,3 @@ class Trait(ABC):
     @abstractmethod
     def apply_ui_to_parameter(self, parameter:Parameter) -> Parameter:
         """Applies UI to the parameter based on the trait."""
-
-
-# This should probably register upon creation
-class TraitRegistry(SingletonMixin):
-    # I'm going to create a dictionary that stores all of the created traits we have so far?
-    # Traits will be associated with certain key words
-    key_to_trait:dict[str, list[Trait.__class__]] = {}
-
-    @classmethod
-    def create_traits(cls, key_word:str) -> list[Trait]|None:
-        if key_word not in cls().key_to_trait:
-             return None
-        values = cls().key_to_trait[key_word]
-        return [trait() for trait in values]
-
-
-    @classmethod
-    def register_trait(cls, trait:Trait) -> None:
-        key_words = trait.get_trait_keys()
-        for key in key_words:
-            if key in cls.key_to_trait:
-                cls().key_to_trait[key].append(trait.__class__)
-            else:
-             cls().key_to_trait[key] = [trait.__class__]
