+++ traits/traits.json
@@ -1,0 +1,4 @@
+{
+    "class_name": "MinMax",
+    "file_name": "minmax.py"
+}
